pyproject.toml
[project]
name = "poc-public-review-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==0.3.25",
    "langchain-openai==0.3.23",
    "langfuse==2.60.8",
    "langgraph==0.4.8",
    "python-dotenv==1.1.0",
    "langchain-anthropic>=0.3.15",
    "langchain-google-genai>=2.1.5",
    "json-repair>=0.47.3",
]
